requirements.txt
gdt-core
gdt-fermi
numpy
scipy
matplotlib
astropy
platformdirs
PyQt6
